Rename common files
