+++ test/onnx/debug_embed_params.py
@@ -1,7 +1,7 @@
 import sys
 
 import onnx
-from test_pytorch_common import flatten
+from pytorch_test_common import flatten
 
 import caffe2.python.onnx.backend as c2
 import torch
