+++ test/onnx/test_models_onnxruntime.py
@@ -7,10 +7,10 @@ from typing import List, Mapping, Tuple
 
 import parameterized
 import PIL
-import test_onnx_common
+import onnx_test_common
 import torchvision
 from test_models import TestModels
-from test_pytorch_common import (
+from pytorch_test_common import (
     TestCase,
     run_tests,
     skipIfUnsupportedMinOpsetVersion,
@@ -37,13 +37,13 @@ def exportTest(self, model, inputs, rtol=1e-2, atol=1e-7, opset_versions=None):
     for opset_version in opset_versions:
         self.opset_version = opset_version
         self.onnx_shape_inference = True
-        test_onnx_common.run_model_test(
+        onnx_test_common.run_model_test(
             self, model, input_args=inputs, rtol=rtol, atol=atol
         )
 
         if self.is_script_test_enabled and opset_version > 11:
             script_model = torch.jit.script(model)
-            test_onnx_common.run_model_test(
+            onnx_test_common.run_model_test(
                 self, script_model, input_args=inputs, rtol=rtol, atol=atol
             )
 
@@ -185,9 +185,9 @@ def _init_test_roi_heads_faster_rcnn():
 @parameterized.parameterized_class(
     ("is_script",),
     ([True, False],),
-    class_name_func=test_onnx_common.parameterize_class_name,
+    class_name_func=onnx_test_common.parameterize_class_name,
 )
-class TestModelsONNXRuntime(test_onnx_common._TestONNXRuntime):
+class TestModelsONNXRuntime(onnx_test_common._TestONNXRuntime):
     @skipIfUnsupportedMinOpsetVersion(11)
     @skipScriptTest()  # Faster RCNN model is not scriptable
     def test_faster_rcnn(self):
