+++ test/onnx/test_pytorch_onnx_caffe2_quantized.py
@@ -8,7 +8,7 @@ import caffe2.python.onnx.backend as c2
 import torch.nn as nn
 import torch.nn.quantized as nnq
 import torch.onnx
-from test_pytorch_common import TestCase, run_tests
+from pytorch_test_common import TestCase, run_tests
 
 
 class TestQuantizedOps(TestCase):
